fix: watch current prop change
