+++ packages/vue/site/App.vue
@@ -10,7 +10,6 @@ import {
     Col, 
     RadioButton, 
     RadioGroup, 
-    Steps, Step, 
     Tree as CTree, 
     TreeNode as CTreeNode, 
     Layout, 
@@ -29,6 +28,7 @@ import StoryTable from './src/story-table.vue';
 import StorySpace from './src/story-space.vue';
 import ThemePicker from './src/theme-picker.vue';
 import StoryVariantUtilities from './src/story-variant-classes.vue';
+import StoryStep from './src/story-step.vue';
 
 const select = ref(null);
 
@@ -39,16 +39,6 @@ const state = ref('sm');
 const menuChangeHandler = () => {
     // eslint-disable-next-line no-console
     console.log(11);
-};
-
-const stepClickable = ref(true);
-const changclick = (step: any) => {
-    if (step.value === 3) {
-        stepClickable.value = false;
-    }
-    else if (step.value < 2 && !stepClickable.value) {
-        stepClickable.value = true;
-    }
 };
 
 const treedata = [
@@ -242,25 +232,7 @@ function log(msg: any) {
                 </Col>
             </Row>
 
-            <div class="step">
-                <Steps :current="1" @change="changclick">
-                    <Step :title="'测试title'" :description="'描述描述描述'" />
-                    <Step :title="'测试title2'" :description="'描述描述描述1'" :current-status="'error'" />
-                    <Step :title="'测试title3'" :description="'描述描述描述2'" :click-able="stepClickable" />
-                    <Step :title="'测试title4'" :description="'描述描述描述3'" />
-                </Steps>
-
-                <Steps>
-                    <Step :title="'测试title'" />
-                    <Step />
-                    <Step>
-                        <template #icon="props">
-                            <i-cosmic-caution :class="props.class" />
-                        </template>
-                    </Step>
-                    <Step :title="'测试title4'" />
-                </Steps>
-            </div>
+            <story-step />
 
             <story-table />
